+++ Makefile.am
@@ -7,28 +7,31 @@ bin_PROGRAMS        = client slave master eclipse cacheserver
 
 # libs -----
 lib_LTLIBRARIES     = libecfs.la
-libecfs_la_SOURCES  = src/common/hash.cc src/common/settings.cc src/common/file_connclient.cc \
-											src/common/fileclient.cc src/common/msgaggregator.cc src/common/util.cc src/common/mapreduce.cc
-libecfs_la_CXXFLAGS = $(AM_CXXFLAGS) -shared
+libecfs_la_SOURCES  = src/common/hash.cc src/common/settings.cc \
+											src/common/fileclient.cc src/common/msgaggregator.cc src/common/util.cc src/common/mapreduce.cc \
+											src/common/dataentry.cc src/common/datablock.cc src/common/messagebuffer.cc
+
+libecfs_la_CXXFLAGS = $(AM_CXXFLAGS) # -shared
 libecfs_la_LDFLAGS  = -version-info 0:0:0
 
 if FS_IS_DHT
-client_SOURCES      = src/client/dht/client.cc src/common/settings.cc
+client_SOURCES      = src/client/dht/client.cc
 client_LDADD        = libecfs.la
-slave_SOURCES       = src/slave/dht/slave.cc src/common/settings.cc
-master_SOURCES      = src/master/dht/master.cc src/common/settings.cc src/common/hash.cc
+slave_SOURCES       = src/slave/dht/slave.cc 
+slave_LDADD         = libecfs.la
+master_SOURCES      = src/master/dht/master.cc 
+master_LDADD        = libecfs.la
 endif
 
-eclipse_SOURCES     = src/orthrus/launcher.cc src/common/settings.cc src/common/hash.cc
+eclipse_SOURCES     = src/file_distributor/launcher.cc
 eclipse_CXXFLAGS    = $(AM_CXXFLAGS) -Wno-maybe-uninitialized
-eclipse_LDDADD      = libecfs.la
-cacheserver_SOURCES = src/orthrus/cacheserver.cc src/common/settings.cc
+eclipse_LDADD       = libecfs.la
+cacheserver_SOURCES = src/cacheserver/cacheserver.cc 
+cacheserver_LDADD   = libecfs.la
 
-
-
-
-pkginclude_HEADERS  = src/common/file_connclient.hh src/common/fileclient.hh \
-                      src/common/hash.hh src/common/msgaggregator.hh src/common/settings.hh
+pkginclude_HEADERS  = src/common/ecfs.hh src/common/file_connclient.hh src/common/fileclient.hh \
+                      src/common/msgaggregator.hh src/common/settings.hh src/common/dataentry.hh \
+											src/common/messagebuffer.hh src/common/datablock.hh src/common/mapreduce.hh
 # Scripts -------
 bin_SCRIPTS         = src/fs/ecfs
 # ------
