--- src/bin/clear_mr.sh
@@ -1,8 +1,0 @@
-rm -f /scratch/youngmoon01/mr_storage/.job*
-
-for line in `cat nodelist.conf`
-do
-	ssh $line rm -f $DHT_PATH/.job* &
-done
-
-wait
