--- src/bin/countline.sh
@@ -1,73 +1,0 @@
-wc -l \
-\
-Makefile \
-countline.sh \
-run.sh \
-run_without_fd.sh \
-clear_mr.sh \
-killall.sh \
-setup.conf \
-\
-bin/fd \
-bin/mrclear \
-bin/mrclear_core \
-bin/mrls \
-bin/mrput \
-bin/mrget \
-bin/mrrm \
-bin/mrcat \
-\
-client/client.cc \
-client/client.hh \
-\
-mapreduce/definitions.hh \
-mapreduce/nfs/mapreduce.hh \
-mapreduce/hdfs/mapreduce.hh \
-mapreduce/dht/mapreduce.hh \
-\
-master/master.cc \
-master/master.hh \
-master/dht/master.cc \
-master/dht/master.hh \
-master/connclient.hh \
-master/connslave.hh \
-master/master_job.hh \
-master/master_task.hh \
-\
-mcc/nfs/mcc.cc \
-mcc/nfs/mcc.hh \
-mcc/hdfs/mcc.cc \
-mcc/hdfs/mcc.hh \
-mcc/dht/mcc.cc \
-mcc/dht/mcc.hh \
-\
-slave/slave.cc \
-slave/slave.hh \
-slave/dht/slave.cc \
-slave/dht/slave.hh \
-slave/slave_job.hh \
-slave/slave_task.hh \
-\
-common/fileclient.hh \
-common/msgaggregator.hh \
-\
-file_distributor/writecount.hh \
-file_distributor/iwriter.hh \
-file_distributor/idistributor.hh \
-file_distributor/ireader.hh \
-file_distributor/fd_core.cc \
-file_distributor/mrcat_core.cc \
-file_distributor/mrrm_core.cc \
-file_distributor/fileserver.hh \
-file_distributor/filebridge.hh \
-file_distributor/filepeer.hh \
-file_distributor/file_connclient.hh \
-file_distributor/messagebuffer.hh \
-\
-orthrus/launcher.cc \
-orthrus/cacheserver.cc \
-orthrus/iwfrequest.hh \
-orthrus/histogram.hh \
-orthrus/datablock.hh \
-orthrus/dataentry.hh \
-orthrus/cache.hh
