--- src/bin/mrcat.in
@@ -1,18 +1,0 @@
-#!/bin/sh
-
-if (( $# < 1 ))
-then
-	echo "mrcat: At least one argument should be provided."
-	exit
-fi
-
-numarg=$#
-
-for (( i=1; i<=$numarg; i++ ))
-do
-	mrcat_core $1
-	sh $MR_PATH/mrcat.sh
-	shift
-done
-
-rm -f $MR_PATH/mrcat.sh
