--- src/bin/mrclear.in
@@ -1,9 +1,0 @@
-#!/bin/sh
-
-NODELIST=$(cat @sysconfdirfull@/eclipse.json | @prefix@/bin/jsawk 'return this.nodes.join("\n")')
-for line in ${NODELIST[@]}
-do
-	ssh $line mrclear_core &
-done
-
-wait
