--- src/bin/mrclear_core
@@ -1,8 +1,0 @@
-#!/bin/sh
-for file in `ls -a $DHT_PATH`
-do
-	if [[ $file != hostname && $file != . && $file != .. ]]
-	then
-		rm -f $DHT_PATH/$file
-	fi
-done
