--- src/bin/mrget.in
@@ -1,15 +1,0 @@
-#!/bin/sh
-
-if (( $# < 1 ))
-then
-	echo "mrget: At least one argument should be provided."
-	exit
-fi
-
-numarg=$#
-
-for (( i=1; i<=$numarg; i++ ))
-do
-	mrcat $1 > $1
-	shift
-done
