--- src/bin/mrls.in
@@ -1,37 +1,0 @@
-#!/bin/sh
-
-rm $MR_PATH/.mrls/* > $MR_PATH/.mrls/tmp 2> $MR_PATH/.mrls/tmp
-NODELIST=$(cat @sysconfdirfull@/eclipse.json | @prefix@/bin/jsawk 'return this.nodes.join("\n")')
-
-i=1
-for line in ${NODELIST[@]}
-do
-	if (( $# > 0 ))
-	then
-		if [ $1 == "all" ]
-		then
-			ssh $line ls -a $DHT_PATH > $MR_PATH/.mrls/data$i.tmp &
-		elif [ $1 == "l" ]
-		then
-			ssh $line ls -lh $DHT_PATH > $MR_PATH/.mrls/data$i.tmp &
-		else
-			ssh $line ls $DHT_PATH > $MR_PATH/.mrls/data$i.tmp &
-		fi
-	else
-		ssh $line ls $DHT_PATH > $MR_PATH/.mrls/data$i.tmp &
-	fi
-	(( i++ ))
-done
-
-wait
-
-i=1
-for line in ${NODELIST[@]}
-do
-	echo -e "\033[0;32mSlave$i:\033[0m" >> $MR_PATH/.mrls/mrls.tmp
-	cat $MR_PATH/.mrls/data$i.tmp >> $MR_PATH/.mrls/mrls.tmp
-	#echo >> $MR_PATH/.mrls/mrls.tmp
-	(( i++ ))
-done
-
-more $MR_PATH/.mrls/mrls.tmp
