--- src/bin/mrput.in
@@ -1,49 +1,0 @@
-#!/bin/sh
-
-if (( $# < 1 ))
-then
-	echo "mrput: At least one argument should be provided."
-	exit
-fi
-
-rm -f $MR_PATH/.mrput/*
-
-rm -f $MR_PATH/.mrput/inputlist.conf
-rm -f $MR_PATH/.mrput/dist.sh
-
-# make new file list
-echo "Generating file list..."
-
-CURPATH=$(readlink -f .)
-
-numarg=$#
-for (( i=0; i<numarg; i++ ))
-do
-	if [ -d $1 ]
-	then
-		echo $1 is a directory. Skipping it...
-	else
-		echo $1 >> $MR_PATH/.mrput/inputlist.conf
-		echo $CURPATH/$1 >> $MR_PATH/.mrput/inputlist.conf
-	fi
-	shift
-done
-
-echo
-
-echo -e "\033[0;32mDone. Number of files: $i\033[0m"
-
-echo
-
-# make new distribution table
-echo "Generating file distribution table..."
-
-@prefix@/bin/fd_core $MR_PATH/.mrput/inputlist.conf $MR_PATH/.mrput/dist.sh
-
-echo -e "\033[0;32mDone\033[0m"
-
-echo "Distributing the files..."
-
-sh $MR_PATH/.mrput/dist.sh
-
-echo -e "\033[0;32mDone\033[0m"
