--- src/bin/mrrm.in
@@ -1,22 +1,0 @@
-#!/bin/sh
-
-if (( $# < 1 ))
-then
-	echo "mrrm: At least one argument should be provided."
-	exit
-fi
-
-numarg=$#
-
-# skip the hostname
-for (( i=1; i<=$numarg; i++ ))
-do
-	if [ $1 != "hostname" ]
-	then
-		mrrm_core $1
-		sh $MR_PATH/mrrm.sh
-	fi
-	shift
-done
-
-rm -f $MR_PATH/mrrm.sh
