+++ src/client/dht/client.cc
@@ -14,7 +14,7 @@
 #include <arpa/inet.h>
 #include <pthread.h>
 #include <netdb.h>
-#include <mapreduce/definitions.hh>
+#include <common/ecfs.hh>
 #include "../common/settings.hh"
 
 using namespace std;
