+++ src/common/datablock.cc
@@ -1,31 +1,8 @@
-#ifndef __DATABLOCK__
-#define __DATABLOCK__
+#include "datablock.hh"
 
 #include <iostream>
-//#include <common/ecfs.hh>
-#include "../common/ecfs.hh"
-
+#include "ecfs.hh"
 #include <string.h>
-
-using namespace std;
-
-class datablock
-{
-    private:
-        char* data;
-        unsigned size; // block size(BLOCKSIZE) is defined in common/ecfs.hh
-        vector<int> recordindex; // index of start of each record
-        
-    public:
-        datablock();
-        ~datablock();
-        
-        int write_record (string record);   // true when succeeded, false when insufficient capacity
-        bool read_record (unsigned pos, string& record);   // -1 when data reaches end of block
-        char* get_data();
-        unsigned get_size();
-        void set_size (unsigned num);
-};
 
 datablock::datablock()
 {
@@ -98,5 +75,3 @@ bool datablock::read_record (unsigned index, string& record)     // return next 
         return false;
     }
 }
-
-#endif
