+++ src/common/hash.cc
@@ -1,4 +1,4 @@
-#include <hash.hh>
+#include "ecfs.hh"
 
 // h {{{
 // -----------------------------------------------
