+++ src/common/msgaggregator.hh
@@ -1,15 +1,16 @@
 #ifndef __MSGAGGREGATOR__
 #define __MSGAGGREGATOR__
 
-#include "ecfs.hh"
-#include <file_distributor/messagebuffer.hh>
-#include <orthrus/dataentry.hh>
 #include <iostream>
 #include <errno.h>
 #include <string.h>
 #include <vector>
 #include <unistd.h>
 #include <stdlib.h>
+
+#include "ecfs.hh"
+#include "messagebuffer.hh"
+#include "dataentry.hh"
 
 using namespace std;
