+++ src/file_distributor/file_connclient.hh
@@ -4,7 +4,7 @@
 #include <iostream>
 #include <fstream>
 #include <fcntl.h>
-#include <orthrus/dataentry.hh>
+#include <common/dataentry.hh>
 #include <file_distributor/filepeer.hh>
 #include "messagebuffer.hh"
 #include "idistributor.hh"
