+++ src/file_distributor/filebridge.hh
@@ -5,11 +5,10 @@
 #include <fstream>
 #include <fcntl.h>
 #include <common/ecfs.hh>
-#include <orthrus/dataentry.hh>
+#include "../common/dataentry.hh"
 #include "filepeer.hh"
 #include "writecount.hh"
 #include "file_connclient.hh"
-#include "../common/settings.hh"
 
 using namespace std;
