+++ src/slave/dht/slave.cc
@@ -13,7 +13,7 @@
 #include <arpa/inet.h>
 #include <netdb.h>
 #include <unistd.h>
-#include <mapreduce/definitions.hh>
+#include <common/ecfs.hh>
 #include "../slave_job.hh"
 #include "../slave_task.hh"
 #include "../common/settings.hh"
